+++ power_sum/classes/PowerInterval.py
@@ -1,5 +1,5 @@
 class PowerInterval:
-    def __init__(self, start: int, end: int, power: float):
+    def __init__(self, start: int, end: int, power: int):
         self.start = start
         self.end = end
         self.power = power
